user 보유 캐시 float -> int 로 롤백

+++ src/main/java/bit/bitgroundspring/util/InitialCashUtil.java
@@ -11,7 +11,7 @@ public class InitialCashUtil {
     @Value("${scheduler.setting}")
     private String schedulerSetting;
     
-    public float getInitialCash() {
+    public int getInitialCash() {
         LocalDateTime now = LocalDateTime.now();
         int day = now.getDayOfMonth();
         int hour = now.getHour();
